Fixed bug in find_lambda function
Thanks to James Brissenden for pointing out this bug, and recommending a fix! 

This bug causes a mistake in the hyperparameter search, s.t. the selected lambda values aren't actually those corresponding to the best loss. 

It is worth noting that this bug was (fortunately) not present in the analyses reported in our paper on TAFKAP. For these analyses, we used a development version of the TAFKAP code which relied on a toolbox by Dimitri Yatsenko for cross validation-based shrinkage estimation of the covariance matrix (https://github.com/atlab/cov-est). When preparing the code to be shared (in a cleaner version that omits the unnecessary baggage and spaghetti code accrued in the development process), we opted to remove the dependence on this toolbox and write our own functions instead. Unfortunately the unit tests we performed on this new code did not reveal the bug we introduced in the process. 

The bug occurred when converting the linear index of the best cross-validated loss for the hyperparameters, to its corresponding coordinates in the grid of possible hyperparameter values. There were actually two mistakes here. First, as James Brissenden pointed out, the previous version was missing an operation best_idx = visited(best_idx), to convert the index within in the list of losses (corresponding only to visited grid locations) to a linear index within the full search grid. Even with this fix, though, there was still a discrepancy between the reported lambda values and what they should have been. This final discrepancy arose on line 453, where we had swapped the dimensions of the grid when converting the linear index back to grid coordinates.

This bug does not affect TAFKAP when run in 'PRINCE-mode' (p.dec_type='PRINCE' ), as PRINCE does not include a hyperparameter search.

We recommend repeating any analyses that you ran with the previous version of TAFKAP, as different hyperparameter values are likely to lead to (somewhat) different results.