Fixed minor bug in basis set selection
A '-1' on line 209 was causing the random basis set index to be drawn from 1:(N-1) instead of 1:N, where N is the desired number of basis sets (i.e. p.precomp_C). Thanks (again) to James Brissenden for spotting this bug. The bug was introduced when adapting our development version of the code (which we used for the analyses in our TAFKAP paper) for wider distribution (i.e. removing unnecessary bloat and cleaning up spaghetti code). The impact of this bug is relatively small (as all basis sets are in principle equally valid), but for optimal results you should still re-run any analyses that included this bug.